добавил экстеншон метод по определению ближайшей даты дня недели
